Remove TieredStorageAssetManager interface
