+++ modules/asset-manager-storage-aws/src/main/java/org/opencastproject/assetmanager/aws/persistence/AwsAssetMappingDto.java
@@ -21,7 +21,7 @@
 
 package org.opencastproject.assetmanager.aws.persistence;
 
-import org.opencastproject.assetmanager.impl.storage.StoragePath;
+import org.opencastproject.assetmanager.api.storage.StoragePath;
 
 import java.util.Date;
 import java.util.List;
