+++ modules/asset-manager-storage-aws/src/test/java/org/opencastproject/assetmanager/aws/persistence/AwsAssetDatabaseImplTest.java
@@ -22,7 +22,7 @@
 package org.opencastproject.assetmanager.aws.persistence;
 
 import org.opencastproject.assetmanager.impl.VersionImpl;
-import org.opencastproject.assetmanager.impl.storage.StoragePath;
+import org.opencastproject.assetmanager.api.storage.StoragePath;
 import org.opencastproject.util.PathSupport;
 import org.opencastproject.util.persistence.PersistenceUtil;
